+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+* dev-master
+    * HOTFIX      #- Fixed executions null variable after deserialize
+
 * 0.2.0 (2016-02-27)
     * ENHANCEMENT #7 Added task execution to hold information for one single try
     * FEATURE     #6 Added cron task